src/main/java/jp/bananafish/ApplicationContextProvider.java
package jp.bananafish;

/**
 * Created by tmikami on 2017/03/11.
 */

import org.springframework.beans.BeansException;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.context.ApplicationContext;
import org.springframework.context.ApplicationContextAware;
import org.springframework.stereotype.Component;

@Component
public class ApplicationContextProvider implements ApplicationContextAware {
    private static ApplicationContext context;

    public static ApplicationContext getApplicationContext() {
        return context;
    }

    public void setApplicationContext(ApplicationContext context) {
        ApplicationContextProvider.context = context;
    }
}
